% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdates.R
\name{cdate_checker}
\alias{cdate_checker}
\title{check censored dates}
\usage{
cdate_checker(x, low.bound, id = NULL)
}
\arguments{
\item{x}{dates, possibly censored, as character or numeric (e.g. 20070101)}

\item{low.bound}{the lower bound}

\item{id}{id variable, must be unique for each x. If NULL 1:n is used}
}
\description{
check censored dates of the type '20110000' or '20110200'.
}
